requirements.txt
Appium-Python-Client~=5.1.1
pytest~=8.4.1
selenium~=4.34.0
openpyxl~=3.1.5
pytest-html~=3.2.0
pytest-xdist~=3.3.1
